feat: 추론 결과 output json형태로 수정 (#66)
